### 2023-06-28 #5600 @editorialbot
[PRE REVIEW]: EPyT: An EPANET-Python Toolkit for Smart Water Network Simulations
**Submitting author:** @mariosmsk (<a href="http://orcid.org/0000-0002-2324-8661">Marios S. Kyriakou</a>)
**Repository:** https://github.com/OpenWaterAnalytics/EPyT/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.6
