+++ paper/paper.bib
@@ -57,6 +57,7 @@ keywords = {Water distribution system, Earthquake, Resilience, Software}
   volume={119},
   pages={710--718},
   year={2015},
+  doi = {10.1016/j.proeng.2015.08.924},
   publisher={Elsevier}
 }
 
