+++ README.md
@@ -2,9 +2,8 @@
 
 [![license](https://img.shields.io/pypi/l/epyt.svg)](https://github.com/KIOS-Research/EPyT/blob/main/LICENSE.md)
 [![pypi](https://img.shields.io/pypi/v/epyt.svg)](https://pypi.org/project/epyt/)
-[![Downloads](https://pepy.tech/badge/epyt)](https://pepy.tech/project/epyt)
-[![Downloads](https://pepy.tech/badge/epyt/month)](https://pepy.tech/project/epyt)
-
+[![Downloads](https://static.pepy.tech/badge/epyt)](https://pepy.tech/project/epyt)
+[![Downloads](https://static.pepy.tech/badge/epyt/month)](https://pepy.tech/project/epyt)
 [![build](https://github.com/OpenWaterAnalytics/EPyT/actions/workflows/build_tests.yml/badge.svg)](https://github.com/OpenWaterAnalytics/EPyT/actions/workflows/build_tests.yml)
 [![Documentation Status](https://readthedocs.org/projects/epanet-python-toolkit-epyt/badge/?version=latest)](https://epanet-python-toolkit-epyt.readthedocs.io/en/latest/?badge=latest)
 
