### 2023-08-23 4b96071
update the function `getNodeTankData` to fix build errors
