### 2023-08-23 59593b6
update the function `getNodeTankData` again / use isinstance instead of type.
