+++ epyt/epanet.py
@@ -2922,7 +2922,7 @@ class epanet:
         lTypes = []
         if len(argv) > 0:
             index = argv[0]
-            if type(index) is list:
+            if isinstance(index, list):
                 for i in index:
                     lTypes.append(self.TYPELINK[self.api.ENgetlinktype(i)])
             else:
@@ -2950,7 +2950,7 @@ class epanet:
         lTypes = []
         if len(argv) > 0:
             index = argv[0]
-            if type(index) is list:
+            if isinstance(index, list):
                 for i in index:
                     lTypes.append(self.api.ENgetlinktype(i))
             else:
@@ -3789,7 +3789,7 @@ class epanet:
         values = []
         if len(argv) > 0:
             index = argv[0]
-            if type(index) is list:
+            if isinstance(index, list):
                 for i in index:
                     values.append(self.api.ENgetlinkindex(i))
             else:
@@ -4133,7 +4133,7 @@ class epanet:
         value = []
         if len(argv) > 0:
             index = argv[0]
-            if type(index) is list:
+            if isinstance(index, list):
                 for i in index:
                     value.append(self.api.ENgetnumdemands(i))
             else:
@@ -4278,7 +4278,7 @@ class epanet:
         values = []
         if len(argv) > 0:
             index = argv[0]
-            if type(index) is list:
+            if isinstance(index, list):
                 for i in index:
                     values.append(self.api.ENgetnodeindex(i))
             else:
@@ -4413,7 +4413,7 @@ class epanet:
         if (len(value) > 0) and (len(argv) > 0):
             index = argv[0]
             try:
-                if type(index) is list:
+                if isinstance(index, list):
                     jIndices = []
                     for i in index:
                         jIndices.append(value[i - 1] + 1)
@@ -4560,7 +4560,7 @@ class epanet:
         if (len(value) > 0) and (len(argv) > 0):
             index = argv[0]
             try:
-                if type(index) is list:
+                if isinstance(index, list):
                     rIndices = []
                     for i in index:
                         rIndices.append(value[i - 1] + 1)
@@ -4668,7 +4668,7 @@ class epanet:
         values = []
         if len(argv) > 0:
             index = argv[0]
-            if type(index) is list:
+            if isinstance(index, list):
                 for i in index:
                     values.append(self.api.ENgetnodeid(i))
             else:
@@ -4912,6 +4912,8 @@ class epanet:
                 result = [True for c in argv[0] if c in tankIndices]
                 if result:
                     tankIndices = argv[0]
+                else:
+                    tankIndices = self.getNodeTankIndex(argv[0])
             else:
                 tankIndices = argv[0]
 
@@ -4966,7 +4968,7 @@ class epanet:
         if (len(value) > 0) and (len(argv) > 0):
             index = argv[0]
             try:
-                if type(index) is list:
+                if isinstance(index, list):
                     tIndices = []
                     for i in index:
                         tIndices.append(value[i - 1] + 1)
@@ -5325,7 +5327,7 @@ class epanet:
         nTypes = []
         if len(argv) > 0:
             index = argv[0]
-            if type(index) is list:
+            if isinstance(index, list):
                 for i in index:
                     nTypes.append(self.TYPENODE[self.api.ENgetnodetype(i)])
             else:
@@ -5348,7 +5350,7 @@ class epanet:
         nTypes = []
         if len(argv) > 0:
             index = argv[0]
-            if type(index) is list:
+            if isinstance(index, list):
                 for i in index:
                     nTypes.append(self.api.ENgetnodetype(i))
             else:
@@ -5774,7 +5776,7 @@ class epanet:
         values = []
         if len(argv) > 0:
             index = argv[0]
[…]
