### 2023-08-23 800ba30
remove testfunctions2, not needed to test with plt ginput for now.
