### 2023-08-23 01672a3
fix docs build by update `html_theme`
