### 2023-10-13 #5600 @TrevorJA
Hi @crvernon, thanks for the invite.  Yes. I've not reviewed for JOSS, but it seems like a good opportunity to start.

I'm pretty constrained these days, but will plan on completing the review by the end of the month.
