### 2023-11-10 #5947 @TrevorJA
Hi @crvernon,  I've made it to the end of the checklist and my review of this package.   I'll share some of my general comments below and close the issue, but please let me know if you need any more info to close out my review.

First off, congrats to @Mariosmsk and the other contributors on putting together a very clean and well designed package.  I've not raised any issues during my review, and recommend the package for JOSS.

**Functionality:**
- I had 0 issues with the install and the usage.  I was able to run each of the 21 example notebooks with no issues. Kudos to the authors for having so many examples.
- The API is very clean.

**Documentation:**
- The authors have formal and automated documentation which is great to see.
- The API would benefit from further categorization of the various functions within `epanet` in addition to the current long list of functions.  I suspect it will be helpful for new users the API had some organization based on use cases or functionality.
- The testing suite is extensive and automated.

 **Software paper:**
 - The paper is well written with a nice select example.  While the authors note that there are a handful of similar software tools to help users connect to EPANET, the EPyT package has the benefit that it "captures the complete function and parameter space of EPANET".  I also appreciate the comparability with the EPANET-MATLAB Toolkit.
 - This package provides a great open-source alternative to an existing Toolkit which requires proprietary programming (MATLAB).
 - The provided use case example in the paper - Monte Carlo based leak detection - gives a nice demonstration of the value of this tool.

Thanks for including me in the review.  Take care!
