### 2023-11-12 #5947 @antonia-had
Hello @crvernon, I am also just done with my checklist and review of this package. Similar to @TrevorJA , I was impressed with the thorough documentation and the multitude of examples provided. I tested the package both in the Terminal (MacOS) and in PyCharm.

Since @TrevorJA mentioned he tried the notebooks, I tested the python scripts and encountered no issues. Some of the examples, e.g., `Plt_EX1_Create_Flow_gif.py`, require additional package installations, which are made clear in the script itself, but maybe it should be mentioned at the top level as "optional" installations for further analysis?

I was also able to replicate the analysis in the paper.

The paper is well written and clear, so is the repository, which made it very easy to review this submission (sorry I took so long)

Minor: In the minimum example provided in the readme, you might want to remove the `>>>` from the script to make it easy to copy and paste

I was also unsure about what the **Automated tests** point in the checklist meant, so I did not check it, but let me know if there's anything else you'd like me to do.

Lastly, this is unorthodox for a publication review, but nice to meet you @Mariosmsk! I'll be visiting Cyprus for a couple weeks in Spring, I might reach out to come visit you all at KIOS and learn more about the cool work you're doing.
