### 2023-11-13 #5947 @crvernon
Great reviews from both @TrevorJA and @antonia-had !  Thanks for the feedback.  Reviews like this make my job easy 🙂

@antonia-had in reference to the **Automated Tests** part, I believe the author satisfies this requirements by having a test suite accompanying the package by which the user can run the tests either locally as well as having a GitHub action setup to run the tests on push to the remote.  An example of not meeting this requirement would be if a author did not 1) have any tests written and included as code with their package, or 2) had text explaining how to test their code but no actual code present to do so.

Thanks so much for your great and timely reviews!
