### 2023-11-13 #5947 @crvernon
@Mariosmsk you have done a fantastic job on this submission!  Thank you for taking care of so much up-front so that the review moves through so easily.

I think the only thing I see that we need to alter is the length of the paper.  JOSS usually requires a 250-1000 word limit on the paper.  Honestly, I hate to do this for your paper because I found it very relevant.  However, let's do try to reduce it a bit to be in keeping with the journal requirements.  I think you can cut the use case and conclusion sections.  You could also cut the target audience section.

If you haven't already, you could put the content you cut in the content of the documentation.

Let's see where we are after that on word count.
