### 2023-11-28 cbd8658
Update paper - R1
Based on JOSS reviewers and editor comments
