+++ paper/paper.md
@@ -64,10 +64,6 @@ Both methods (direct calls and wrappers) are limited to the functions made avail
 3.	Provide code templates for expanding the capabilities of EPANET and facilitating smart water systems research, while supporting the adoption of open-science and reproducible research best practices.
 4.	Provide a similar data structure in Python and MATLAB, to facilitate translation between the two environments.
 
-## Target audience
-
-The target audience is students and researchers in engineering (civil, environmental, chemical, mechanical, electrical) and computer science, who want to utilize data analytics and Artificial Intelligence (AI) methods in applications related to drinking water networks. Moreover, the toolkit can be used by researchers, data scientists, and engineers working in startups and companies for the development of new innovative smart water solutions.
-
 ## State of the field
 
 As a precursor of EPyT, the open-source [EPANET-MATLAB Toolkit](https://github.
@@ -79,7 +75,7 @@ the *OpenWaterAnalytics Community*. Moreover, EMT introduced the `epanet` class,
 structures and function names which are human-readable and self-explanatory. In addition to EPyT, a few other 
 relevant Python-based tools connect with EPANET. The most advanced, and relevant, is the [Water Network Tool for Resilience](https://github.com/USEPA/WNTR)  (WNTR), which was developed by the US EPA and Sandia National Laboratories, and released under the Revised BSD license [@klise2017]. `WNTR` facilitates the simulation of both hydraulic and quality dynamics, and in addition, it allows the simulation of various events such as pipe breaks, disasters such as earthquakes, power outages, fires, and contamination events. At the moment, `WNTR` includes only a subset of EPANET functions necessary for its simulation capabilities. The [Object-Oriented Pipe Network Analyzer](https://github.com/oopnet/oopnet) (OOPNET) is a Python package that allows modelling and simulating hydraulic water distribution systems  [@steffelbauer2015]. A drawback is that, since `OOPNET` is based on the runtime executable of EPANET, it does not provide access to the internal library functions. The [OWA-EPANET](https://pypi.org/project/owa-epanet/) is a SWIG auto-generated "thin" wrapper around the EPANET libraries. The goal of this package is to provide a Python interface that requires minimal effort to keep up to date with the core library and can be used by higher-level applications. Other Python-based EPANET toolkits include [epynet](https://github.com/Vitens/epynet), developed by Vitens, and [epanettools](https://pypi.org/project/EPANETTOOLS/) which supports older versions of the EPANET toolkit. Recently a new tool has been published, [viswaternet](https://pypi.org/project/viswaternet/) which provides a tool for visualizing static and time-varying attributes of EPANET-based water distribution systems; this tool can be used in parallel with EPyT for visualization purposes [@thomas2023].
 
-A key unique feature of EPyT, is that it captures the complete function and parameter space of EPANET. Another important aspect of EPyT is that it shares the same function names as the EPANET-MATLAB Toolkit. Our motivation is that this will facilitate the transition of state-of-the-art code originating from academia (which typically uses MATLAB) to more industrial applications, which typically use Python due to its open-source license and its extended set of data analytics modules. 
[…]
