+++ docs/source/conf.py
@@ -26,7 +26,7 @@ sys.path.insert(0, os.path.join(pathlib.Path(__file__).parents[2].resolve().as_p
 author = 'KIOS CoE'
 
 # The full version, including alpha/beta/rc tags
-release = '1.0.5'
+release = '1.0.7'
 
 # -- General configuration ---------------------------------------------------
 
