+++ paper/paper.md
@@ -78,7 +78,7 @@ A key unique feature of EPyT, is that it captures the complete function and para
 
 # Functionality
 The EPyT python class, `epanet`, includes properties of the input network model, static properties, public 
-functions, and local functions that directly call the EPANET shared object library or the executable. The full API with detailed examples is provided in the [link](https://epanet-python-toolkit-epyt.readthedocs.io/en/latest/api.html). In the following paragraphs, we introduce a selection of the capabilities of the tool and an illustrative smart water research example.
+functions, and local functions that directly call the EPANET shared object library or the executable. The full API with detailed examples is provided in the [link](https://epanet-python-toolkit-epyt.readthedocs.io/en/stable/api.html). In the following paragraphs, we introduce a selection of the capabilities of the tool and an illustrative smart water research example.
 
 Through EPyT, users can easily load EPANET network models, as demonstrated by importing the module and loading the 
 `L-TOWN.inp` network in a provided example [@vrachimis2022].
