### 2023-11-28 2419711
Update examples.rst
