### 2023-11-28 33cd9f5
Update examples.rst
