+++ docs/source/examples.rst
@@ -46,598 +46,7 @@ Simple example
      'CurveCount',
      'CurveIndex',
      'CurvesInfo',
-     'DEMANDMODEL',
-     'InputFile',
-     'Iterations',
-     'LOGOP',
-     'LibEPANET',
-     'LibEPANETpath',
-     'LinkBulkReactionCoeff',
-     'LinkCount',
-     'LinkDiameter',
-     'LinkFlowUnits',
-     'LinkIndex',
-     'LinkInitialSetting',
-     'LinkInitialStatus',
-     'LinkLength',
-     'LinkMinorLossCoeff',
-     'LinkNameID',
-     'LinkPipeCount',
-     'LinkPipeIndex',
-     'LinkPipeNameID',
-     'LinkPumpCount',
-     'LinkPumpHeadCurveIndex',
-     'LinkPumpIndex',
-     'LinkPumpNameID',
-     'LinkPumpPatternIndex',
-     'LinkPumpPatternNameID',
-     'LinkPumpPower',
-     'LinkPumpType',
-     'LinkPumpTypeCode',
-     'LinkRoughnessCoeff',
-     'LinkType',
-     'LinkTypeIndex',
-     'LinkValveCount',
-     'LinkValveIndex',
-     'LinkValveNameID',
-     'LinkWallReactionCoeff',
-     'NodeBaseDemands',
-     'NodeCoordinates',
-     'NodeCount',
-     'NodeDemandCategoriesNumber',
-     'NodeDemandPatternIndex',
-     'NodeDemandPatternNameID',
-     'NodeElevations',
-     'NodeEmitterCoeff',
-     'NodeIndex',
-     'NodeInitialQuality',
-     'NodeJunctionCount',
-     'NodeJunctionIndex',
-     'NodeJunctionNameID',
-     'NodeNameID',
-     'NodePatternIndex',
-     'NodeReservoirCount',
-     'NodeReservoirIndex',
-     'NodeReservoirNameID',
-     'NodeSourcePatternIndex',
-     'NodeSourceQuality',
-     'NodeSourceTypeIndex',
-     'NodeTankBulkReactionCoeff',
-     'NodeTankCount',
-     'NodeTankDiameter',
-     'NodeTankIndex',
-     'NodeTankInitialLevel',
-     'NodeTankInitialWaterVolume',
-     'NodeTankMaximumWaterLevel',
-     'NodeTankMaximumWaterVolume',
-     'NodeTankMinimumFraction',
-     'NodeTankMinimumWaterLevel',
-     'NodeTankMinimumWaterVolume',
-     'NodeTankMixZoneVolume',
-     'NodeTankMixingModelCode',
-     'NodeTankMixingModelType',
-     'NodeTankNameID',
-     'NodeTankReservoirCount',
-     'NodeTankVolumeCurveIndex',
-     'NodeType',
-     'NodeTypeIndex',
-     'NodesConnectingLinksID',
-     'NodesConnectingLinksIndex',
-     'OptionsAccuracyValue',
-     'OptionsEmitterExponent',
-     'OptionsFlowChange',
-     'OptionsHeadError',
-     'OptionsHeadLossFormula',
-     'OptionsHydraulics',
-     'OptionsMaxTrials',
-     'OptionsPattern',
-     'OptionsPatternDemandMultiplier',
-     'OptionsQualityTolerance',
-     'OptionsSpecificGravity',
-     'OptionsUnbalanced',
-     'OptionsViscosity',
-     'Pattern',
-     'PatternAverageValue',
-     'PatternCount',
-     'PatternIndex',
-     'PatternLengths',
-     'PatternNameID',
-     'QualityChemName',
[…]
