### 2023-11-28 3cfc8e2
added text to the examples
