+++ docs/api.rst
@@ -1,594 +1,0 @@
-API documentation
------------------
-
-.. currentmodule:: epanet
-
-.. autoclass:: epanet
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~epanet.__init__
-      ~epanet.addControls
-      ~epanet.addCurve
-      ~epanet.addLinkPipe
-      ~epanet.addLinkPipeCV
-      ~epanet.addLinkPump
-      ~epanet.addLinkValveFCV
-      ~epanet.addLinkValveGPV
-      ~epanet.addLinkValvePBV
-      ~epanet.addLinkValvePRV
-      ~epanet.addLinkValvePSV
-      ~epanet.addLinkValveTCV
-      ~epanet.addNodeJunction
-      ~epanet.addNodeJunctionDemand
-      ~epanet.addNodeReservoir
-      ~epanet.addNodeTank
-      ~epanet.addPattern
-      ~epanet.addRules
-      ~epanet.appRotateNetwork
-      ~epanet.appShiftNetwork
-      ~epanet.arange
-      ~epanet.clearReport
-      ~epanet.closeHydraulicAnalysis
-      ~epanet.closeNetwork
-      ~epanet.closeQualityAnalysis
-      ~epanet.copyReport
-      ~epanet.createProject
-      ~epanet.deleteAllTemps
-      ~epanet.deleteControls
-      ~epanet.deleteCurve
-      ~epanet.deleteLink
-      ~epanet.deleteNode
-      ~epanet.deleteNodeJunctionDemand
-      ~epanet.deletePattern
-      ~epanet.deleteProject
-      ~epanet.deleteRules
-      ~epanet.getCMDCODE
-      ~epanet.getComputedHydraulicTimeSeries
-      ~epanet.getComputedQualityTimeSeries
-      ~epanet.getComputedTimeSeries
-      ~epanet.getComputedTimeSeries_ENepanet
-      ~epanet.getConnectivityMatrix
-      ~epanet.getControlRulesCount
-      ~epanet.getControls
-      ~epanet.getCounts
-      ~epanet.getCurveComment
-      ~epanet.getCurveCount
-      ~epanet.getCurveIndex
-      ~epanet.getCurveLengths
-      ~epanet.getCurveNameID
-      ~epanet.getCurveType
-      ~epanet.getCurveTypeIndex
-      ~epanet.getCurveValue
-      ~epanet.getCurvesInfo
-      ~epanet.getDemandModel
-      ~epanet.getENfunctionsImpemented
-      ~epanet.getError
-      ~epanet.getFlowUnits
-      ~epanet.getLibFunctions
-      ~epanet.getLinkActualQuality
-      ~epanet.getLinkBulkReactionCoeff
-      ~epanet.getLinkComment
-      ~epanet.getLinkCount
-      ~epanet.getLinkDiameter
-      ~epanet.getLinkEnergy
-      ~epanet.getLinkFlows
-      ~epanet.getLinkHeadloss
-      ~epanet.getLinkIndex
-      ~epanet.getLinkInitialSetting
-      ~epanet.getLinkInitialStatus
-      ~epanet.getLinkLength
-      ~epanet.getLinkMinorLossCoeff
-      ~epanet.getLinkNameID
-      ~epanet.getLinkNodesIndex
-      ~epanet.getLinkPipeCount
-      ~epanet.getLinkPipeIndex
-      ~epanet.getLinkPipeNameID
-      ~epanet.getLinkPumpCount
-      ~epanet.getLinkPumpECost
[…]
