+++ docs/index.rst
@@ -21,8 +21,8 @@
         :alt: Documentation Status
 
 
-Contents
---------
+Welcome to EPyT's documentation!
+================================
 
 .. toctree::
    :maxdepth: 2
@@ -33,3 +33,12 @@ Contents
    source/api
    examples
    changelog
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
