+++ docs/modules.rst
@@ -1,0 +1,8 @@
+EPyT
+====
+
+.. toctree::
+   :maxdepth: 4
+
+   epyt
+   setup
