+++ docs/epyt.tests.rst
@@ -1,0 +1,61 @@
+epyt.tests package
+==================
+
+Submodules
+----------
+
+epyt.tests.general\_unit\_test module
+-------------------------------------
+
+.. automodule:: epyt.tests.general_unit_test
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+epyt.tests.net1\_unit\_test module
+----------------------------------
+
+.. automodule:: epyt.tests.net1_unit_test
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+epyt.tests.testFunctions module
+-------------------------------
+
+.. automodule:: epyt.tests.testFunctions
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+epyt.tests.testFunctions\_ex module
+-----------------------------------
+
+.. automodule:: epyt.tests.testFunctions_ex
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+epyt.tests.testGetSetLinks module
+---------------------------------
+
+.. automodule:: epyt.tests.testGetSetLinks
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+epyt.tests.testGetSetNodes module
+---------------------------------
+
+.. automodule:: epyt.tests.testGetSetNodes
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: epyt.tests
+   :members:
+   :undoc-members:
+   :show-inheritance:
