+++ docs/epyt.examples.rst
@@ -1,0 +1,10 @@
+epyt.examples package
+=====================
+
+Module contents
+---------------
+
+.. automodule:: epyt.examples
+   :members:
+   :undoc-members:
+   :show-inheritance:
