+++ docs/epyt.networks.rst
@@ -1,0 +1,10 @@
+epyt.networks package
+=====================
+
+Module contents
+---------------
+
+.. automodule:: epyt.networks
+   :members:
+   :undoc-members:
+   :show-inheritance:
