+++ docs/epyt.rst
@@ -1,0 +1,32 @@
+epyt package
+============
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   epyt.examples
+   epyt.libraries
+   epyt.networks
+   epyt.tests
+
+Submodules
+----------
+
+epyt.epanet module
+------------------
+
+.. automodule:: epyt.epanet
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: epyt
+   :members:
+   :undoc-members:
+   :show-inheritance:
