+++ docs/requirements.txt
@@ -1,4 +1,5 @@
 sphinx-copybutton
+sphinx_rtd_theme
 numpy
 matplotlib
 pandas
