+++ docs/usage.rst
@@ -5,16 +5,16 @@ Minimum Example:
 
 .. code-block:: console
    
-   >>> from epyt import epanet
-   >>> d = epanet('Net1.inp')
-   >>> d.getNodeCount()
-   >>> d.getNodeElevations()
+   from epyt import epanet
+   d = epanet('Net1.inp')
+   d.getNodeCount()
+   d.getNodeElevations()
 
 Plot the network:
 
 .. code-block:: console
 
-   >>> d.plot()
+   d.plot()
 
 .. image:: net1_plot.png
         :target: net1_plot.png
@@ -23,13 +23,13 @@ Lists all available functions and properties:
 
 .. code-block:: console
 
-   >>> dir(d)
+   dir(d)
 
 Retrieve some examples for the function:
 
 .. code-block:: console
 
-   >>> help(d.getNodeElevations)
+   help(d.getNodeElevations)
 
     Help on method getNodeElevations in module epyt.epanet:
     getNodeElevations(*argv) method of epyt.epanet.epanet instance
