+++ docs/epyt.rst
@@ -1,20 +1,6 @@
 epyt package
 ============
 
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   epyt.examples
-   epyt.libraries
-   epyt.networks
-   epyt.tests
-
-Submodules
-----------
-
 epyt.epanet module
 ------------------
 
