+++ docs/epyt.examples.rst
@@ -1,10 +1,0 @@
-epyt.examples package
-=====================
-
-Module contents
----------------
-
-.. automodule:: epyt.examples
-   :members:
-   :undoc-members:
-   :show-inheritance:
