+++ docs/epyt.libraries.rst
@@ -1,10 +1,0 @@
-epyt.libraries package
-======================
-
-Module contents
----------------
-
-.. automodule:: epyt.libraries
-   :members:
-   :undoc-members:
-   :show-inheritance:
