+++ docs/api.rst
@@ -1,18 +1,7 @@
-epyt package
-============
+API Documentation
+=================
 
-epyt.epanet module
-------------------
-
-.. automodule:: epyt.epanet
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: epyt
+.. automodule:: epyt.epanet.epanet
    :members:
    :undoc-members:
    :show-inheritance:
