### 2023-11-29 #5947 @crvernon
:wave: @Mariosmsk thanks for making the paper changes.  Even though we are still over the word count, I see you were able to bring it down by around 700 words.  I read over the content and I believe everything in there is appropriate.  I will let the EiC make the final call but I am going to move forward.

So here is what we have left to do:

- [ ] Conduct a GitHub release of the current reviewed version of the software you now have on the main and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)

- [ ] Please list the DOI of the archived version here
