+++ epyt/__init__.py
@@ -1,10 +1,9 @@
 # -*- coding: utf-8 -*-
-__author__ = """KIOS CoE"""
+__author__ = """Marios S. Kyriakou"""
 __email__ = "kiriakou.marios@ucy.ac.cy"
-__version__ = "1.0.7"
+__version__ = "1.0.8"
 __copyright__ = """Copyright 2022, KIOS Research and Innovation Center of Excellence (KIOS CoE), 
 University of Cyprus (www.kios.org.cy)."""
 __license__ = "EUPL License, Version 1.2"
-
 
 from epyt.epanet import epanet
