### 2023-11-30 #5947 @crvernon
❗  **Note for EiC:**

Please see comments on the current paper length in the following:

**Here**:

> @Mariosmsk you have done a fantastic job on this submission! Thank you for taking care of so much up-front so that the review moves through so easily.
>
> I think the only thing I see that we need to alter is the length of the paper. JOSS usually requires a 250-1000 word limit on the paper. Honestly, I hate to do this for your paper because I found it very relevant. However, let's do try to reduce it a bit to be in keeping with the journal requirements. I think you can cut the use case and conclusion sections. You could also cut the target audience section.
>
> If you haven't already, you could put the content you cut in the content of the documentation.
>
> Let's see where we are after that on word count.

**and as part of a follow-up comment here:**

> 👋 @Mariosmsk thanks for making the paper changes. Even though we are still over the word count, I see you were able to bring it down by around 700 words. I read over the content and I believe everything in there is appropriate. I will let the EiC make the final call but I am going to move forward.

As mentioned in the comments above, the word count of the paper is still higher than usual.  However, I found the content relevant to the work since the authors are providing an extension and novel Python wrapper to an existing widely used software product.  If I were a user of the software represented in this paper, I would want the description provided by the authors as context.  I do understand that the word count exceeds our normal bounds so I default to your opinion on whether or not further word count reduction should be required.

Thanks!
